Export Wh as INT to prevent issues when export 0 (zero)
When the db contains 0, influx does not accept a float(0) as value for Wh
Since these values are always integer, changed it to INT form both daily as details

+++ app/models/export_influxdb.py
@@ -40,7 +40,7 @@ class ExportInfluxDB:
                     "month": daily.date.strftime("%m"),
                 },
                 fields={
-                    "Wh": float(watt),
+                    "Wh": int(watt),
                     "kWh": float(forceRound(kwatt, 5)),
                     "price": float(forceRound(euro, 5))
                 },
@@ -73,7 +73,7 @@ class ExportInfluxDB:
                     "measure_type": detail.measure_type,
                 },
                 fields={
-                    "W": float(watt),
+                    "W": int(watt),
                     "kW": float(forceRound(kwatt, 5)),
                     "Wh": float(watth),
                     "kWh": float(forceRound(kwatth, 5)),
